Add files via upload
Write a program that uses following list of cities per country,
usa = [ “atlanta”, “new york”, “chicago”, “baltimore”]
uk = [“london”, “bristol”, “cambridge”]
india = [“mumbai”,”delhi”, “banglore”]
a. Write a program that asks user to enter a city name and it should tell you which
country it belongs to
b. Write a program that asks user to enter two cities and tell you if they both are in
same country or not